requirements.txt
datasets
fastapi
huggingface-hub
matplotlib
numpy
pandas
pillow
pydantic
# torch
torchvision
tqdm
transformers==4.48.0
vllm
wandb
peft
einops 
timm
albumentations
torchmetrics
openai
anthropic
pymupdf
pybind11
joblib
tenacity
bitsandbytes